# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nitip Core REST API server.",
        "title": "Nitip Core API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@nitip.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/trips": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Declare a new journey as a runner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] Trip"
                ],
                "summary": "Create a new trip",
                "parameters": [
                    {
                        "description": "Trip payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_trip.CreateTripRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_domain_trip.Trip"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/trips/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all trips declared by the logged-in runner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] Trip"
                ],
                "summary": "Get runner trips",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_domain_trip.Trip"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/trips/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a trip as cancelled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] Trip"
                ],
                "summary": "Cancel a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/trips/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a started trip as completed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] Trip"
                ],
                "summary": "Complete a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/trips/{id}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a trip as started (in progress)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Runner] Trip"
                ],
                "summary": "Start a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_codecoffy_nitip-core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_domain_trip.CreateTripRequest": {
            "type": "object",
            "required": [
                "departure_time",
                "destination_lat",
                "destination_lng",
                "destination_name",
                "origin_lat",
                "origin_lng",
                "origin_name",
                "vehicle_type"
            ],
            "properties": {
                "allowed_service_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "departure_time": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lng": {
                    "type": "number"
                },
                "destination_name": {
                    "type": "string"
                },
                "is_round_trip": {
                    "type": "boolean"
                },
                "max_volume_liters": {
                    "type": "number"
                },
                "max_weight_kg": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "origin_lat": {
                    "type": "number"
                },
                "origin_lng": {
                    "type": "number"
                },
                "origin_name": {
                    "type": "string"
                },
                "return_time": {
                    "type": "string"
                },
                "vehicle_type": {
                    "type": "string",
                    "enum": [
                        "motorcycle",
                        "car",
                        "pickup"
                    ]
                }
            }
        },
        "internal_domain_trip.Trip": {
            "type": "object",
            "properties": {
                "allowed_service_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available_volume_liters": {
                    "type": "number"
                },
                "available_weight_kg": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lng": {
                    "type": "number"
                },
                "destination_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_round_trip": {
                    "type": "boolean"
                },
                "max_volume_liters": {
                    "type": "number"
                },
                "max_weight_kg": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "origin_lat": {
                    "type": "number"
                },
                "origin_lng": {
                    "type": "number"
                },
                "origin_name": {
                    "type": "string"
                },
                "return_time": {
                    "type": "string"
                },
                "runner_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vehicle_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}